pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "academie-fr-dnpd-tgbot"
version = "0.1.0"
description = "Telegram bot: envoi quotidien 'Dire, ne pas dire' (Académie française) sans doublons."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot==21.6",
  "APScheduler==3.10.4",
  "requests==2.32.3",
  "beautifulsoup4==4.12.3",
]

[tool.setuptools]
packages = ["bot"]
